fix: create-user isolated, get-user improvements
